fix not copy config solution
