+++ RevitPythonShell/Properties/launchSettings.json
@@ -43,6 +43,14 @@
     "Revit 2021": {
       "commandName": "Executable",
       "executablePath": "%ProgramW6432%\\Autodesk\\Revit 2021\\Revit.exe"
+    }     ,
+    "Revit 2022": {
+      "commandName": "Executable",
+      "executablePath": "%ProgramW6432%\\Autodesk\\Revit 2022\\Revit.exe"
+    }     ,
+    "Revit 2023": {
+      "commandName": "Executable",
+      "executablePath": "%ProgramW6432%\\Autodesk\\Revit 2023\\Revit.exe"
     }
   }
 }
