+++ RevitPythonShell/RevitPythonShell.addin
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RevitAddIns>
+  <AddIn Type="Application">
+    <FullClassName>RevitPythonShell.App</FullClassName>
+    <ClientId>3a7a1d24-51ed-462b-949f-1ddcca12008d</ClientId>
+    <Name>RevitPythonShell</Name>
+    <Assembly>RevitPythonShell/RevitPythonShell.dll</Assembly>
+    <VendorId>3a7a1d24-51ed-462b-949f-1ddcca12008d</VendorId>
+    <VendorDescription>RIPS, RIPS, </VendorDescription>
+  </AddIn>
+</RevitAddIns>
